Server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
jinja2==3.1.2
aiofiles==23.2.1
PyJWT==2.10.1
opencv-python==4.9.0.80
mediapipe==0.10.21
numpy==1.26.4
tensorflow==2.15.0
pandas==2.0.3
hmmlearn==0.3.0 